feat(tl-9fj): extract shared Python logging into services/tl-logging
All 4 Python services (search, ingestion, tutoring-service,
analytics-service) had byte-for-byte duplicate logging_config.py
files.  Extract the canonical implementation to a new pip-installable
package `tl-logging` (tl_logging.logging_config) so future changes
happen in one place.

Each service's app/logging_config.py is now a thin re-export shim:
  from tl_logging.logging_config import configure_logging, ...

CI changes:
- python-services job: installs ../tl-logging before requirements.txt
- docker-build job: copies tl-logging/ into the service build context
  before docker build (Dockerfiles use COPY tl-logging/ ./tl-logging/)
- path filter: services/tl-logging/** triggers python-services + docker

All 48 logging tests pass across all 4 services.
