+++ services/analytics-service/requirements.txt
@@ -1,3 +1,6 @@
+# Shared structured logging package (canonical source: services/tl-logging/)
+../tl-logging
+
 fastapi==0.115.6
 uvicorn[standard]==0.34.0
 asyncpg==0.30.0
