+++ services/search/Dockerfile
@@ -2,6 +2,8 @@ FROM python:3.12-slim
 
 WORKDIR /app
 
+# tl-logging is copied here by CI before docker build (see ci.yml docker-build job)
+COPY tl-logging/ ./tl-logging/
 COPY requirements.txt .
 RUN pip install --no-cache-dir -r requirements.txt
