+++ services/search/requirements.txt
@@ -1,3 +1,6 @@
+# Shared structured logging package (canonical source: services/tl-logging/)
+../tl-logging
+
 fastapi==0.115.0
 uvicorn[standard]==0.30.6
 pydantic==2.9.2
