+++ services/tutoring-service/requirements.txt
@@ -1,3 +1,6 @@
+# Shared structured logging package (canonical source: services/tl-logging/)
+../tl-logging
+
 fastapi==0.115.6
 uvicorn[standard]==0.42.0
 httpx==0.28.1
